refactor: remove server discovery subsystem
Consumer audit of the two known BCK consumers found zero references to the Bonjour/network discovery subsystem. Removes ~1,643 LOC of unused code that auto-detected Ollama and LM Studio servers on the local network. Apps that need discovery can manually enter server URLs or wire their own scanner.

Deletes:
- BonjourDiscoveryService, NetworkDiscoveryService, ServerDiscoveryService protocol
- DiscoveredServer model, ServerType enum
- ServerDiscoveryView, ServerDiscoveryViewModel
- MockServerDiscoveryService test helper
- showServerDiscovery feature flag
- 3 test files

RemoteModelInfo (the Ollama tag-list response model) is preserved alongside OllamaModelListService since that service has its own test coverage and is the only consumer.

BREAKING CHANGE: ServerDiscoveryService protocol, DiscoveredServer, ServerType, ServerDiscoveryView, and the showServerDiscovery feature flag are removed. Apps that reference them directly will not compile against 0.6.0.
